app: use AwaitQueue to queue server side consumers related requests/notifications
This avoids a potential problem that may happen if another peer creates a Producer, then immediately closes it and then creates another one. In that case, before this change, the consuming peer may get black video or no audio since it would have ignored the `consumerClosed` notification and hence it may render the closed consumer isntead of the new one.